Enforced variable naming conventions.

+++ Assets/Scripts/Entities/Base/EntityData.cs
@@ -8,7 +8,7 @@ public class EntityData : ScriptableObject
     [field: SerializeField]
     public EntityBase Prefab { get; protected set; }
     [field: SerializeField]
-    public float MaxHp { get; protected set; }
+    public int MaxHp { get; protected set; }
     [field: SerializeField]
-    public float MaxStamina { get; protected set; }
+    public int MaxStamina { get; protected set; }
 }
